Initial attempt to refactor part of Optimzation.h into a distinct interface.
Split constraints into a separate header file.

Optimization.h is still longer than I'd like, but at least now some of
the code which doesn't have template arguments has moved into a
distinct compilation unit.  Also MathematicalProgram has stopped
reaching into private data members of a mutable object, so that's
nice.
