+++ drake/solvers/CMakeLists.txt
@@ -33,6 +33,7 @@ if(MATLAB_FOUND)
 endif()
 
 set(optimization_files)
+list(APPEND optimization_files MathematicalProgram.cpp)
 if (snopt_c_FOUND)
   list(APPEND optimization_files SNOPTWrapper.cpp)
 else()
@@ -40,7 +41,7 @@ else()
 endif(snopt_c_FOUND)
 add_library_with_exports(LIB_NAME drakeOptimization SOURCE_FILES ${optimization_files})
 set_target_properties(drakeOptimization PROPERTIES COMPILE_FLAGS -fPIC)
-drake_install_headers(Optimization.h)
+drake_install_headers(MathematicalProgram.h Optimization.h)
 pods_install_libraries(drakeOptimization)
 if (snopt_c_FOUND)
   pods_use_pkg_config_packages(drakeOptimization snopt_c)
